+++ app/src/main/res/layout/registration_sucessful.xml
@@ -63,6 +63,16 @@
                 android:text="@string/get_started"
                 android:textColor="@color/colorAccent" />
 
+            <Button
+                android:id="@+id/view_profile"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="5dp"
+                android:background="@color/colorPrimaryDark"
+                android:fontFamily="@font/arvo"
+                android:text="View Your Profile (test)"
+                android:textColor="@color/colorAccent" />
+
         </LinearLayout>
 
     </LinearLayout>
