+++ app/src/main/res/layout/venue_profile_creation.xml
@@ -264,69 +264,6 @@
             android:gravity="center">
 
             <android.support.design.widget.TextInputEditText
-                android:id="@+id/venue_address_line_2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/address_line_2"
-                android:inputType="text"
-                android:textColor="@color/colorAccent"
-                android:textSize="18sp" />
-
-        </android.support.design.widget.TextInputLayout>
-
-        <android.support.design.widget.TextInputLayout
-            android:layout_marginTop="10dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingLeft="16dp"
-            android:paddingRight="16dp"
-            android:fontFamily="@font/arvo"
-            android:textColorHint="@color/colorAccent"
-            android:gravity="center">
-
-            <android.support.design.widget.TextInputEditText
-                android:id="@+id/venue_address_line_town"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/address_line_town"
-                android:inputType="text"
-                android:textColor="@color/colorAccent"
-                android:textSize="18sp" />
-
-        </android.support.design.widget.TextInputLayout>
-
-        <android.support.design.widget.TextInputLayout
-            android:layout_marginTop="10dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingLeft="16dp"
-            android:paddingRight="16dp"
-            android:fontFamily="@font/arvo"
-            android:textColorHint="@color/colorAccent"
-            android:gravity="center">
-
-            <android.support.design.widget.TextInputEditText
-                android:id="@+id/venue_address_line_county"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/address_line_county"
-                android:inputType="text"
-                android:textColor="@color/colorAccent"
-                android:textSize="18sp" />
-
-        </android.support.design.widget.TextInputLayout>
-
-        <android.support.design.widget.TextInputLayout
-            android:layout_marginTop="10dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingLeft="16dp"
-            android:paddingRight="16dp"
-            android:fontFamily="@font/arvo"
-            android:textColorHint="@color/colorAccent"
-            android:gravity="center">
-
-            <android.support.design.widget.TextInputEditText
                 android:id="@+id/venue_address_line_postcode"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
